% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arbitrary_derivatives.R
\name{get_arbitrary_derivatives}
\alias{get_arbitrary_derivatives}
\title{Get derivatives at arbitrary temperatures}
\usage{
get_arbitrary_derivatives(experiment_folder, experiment_design_filename)
}
\arguments{
\item{experiment_folder}{The folder where the experiment is stored}

\item{experiment_design_filename}{The file with the experiment design}
}
\value{
Data base with the derivatives at arbitrary temperatures
}
\description{
Get derivatives at arbitrary temperatures
}
\examples{
NULL
}
