% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_sum_rel_b_opt.R
\name{get_community_performance_optimum_measures}
\alias{get_community_performance_optimum_measures}
\title{Get realized performance-optimum summaries relative to environmental temperature}
\usage{
get_community_performance_optimum_measures(temperatures, expt)
}
\arguments{
\item{temperatures}{Connection to the database containing temperature data}

\item{expt}{Dataset containing experiment information}
}
\value{
A dataset containing realized performance-optimum summaries relative
to the mean environmental temperature experienced by each community.
}
\description{
Get realized performance-optimum summaries relative to environmental temperature
}
\examples{
NULL
}
