% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_popstab.R
\name{get_community_popstab}
\alias{get_community_popstab}
\title{Get the stability of each case (community). Calculated CV of total biomass temporal variation. Also returns mean total biomass and standard deviation of total biomass}
\usage{
get_community_popstab(dynamics)
}
\arguments{
\item{dynamics}{Connection to database containing dynamics data}
}
\value{
A dataset containing the stability of each case (community). Calculated CV of total biomass temporal variation. Also returns mean total biomass and standard deviation of total biomass
}
\description{
Get the stability of each case (community). Calculated CV of total biomass temporal variation. Also returns mean total biomass and standard deviation of total biomass
}
\examples{
NULL
}
