% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_response_diversity.R
\name{get_community_response_diversity}
\alias{get_community_response_diversity}
\title{Get measures of community response diversity}
\usage{
get_community_response_diversity(temp_derivs)
}
\arguments{
\item{temp_derivs}{Connection to database containing temporal derivatives}
}
\value{
A dataset containing measures of community response diversity
}
\description{
Get measures of community response diversity
}
\examples{
NULL
}
