% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_temp_sens.R
\name{get_community_temp_sens}
\alias{get_community_temp_sens}
\title{Get the sensitivity of total biomass to temperature variation. Currently measured as the slope of a linear regression of total biomass on temperature.}
\usage{
get_community_temp_sens(dynamics, temperatures, rollsumr_window = 50)
}
\arguments{
\item{dynamics}{Connection to database containing dynamics data}

\item{temperatures}{Connection to database containing temperature data}

\item{rollsumr_window}{The window size for the rolling sum of temperature. Default is 50}

\item{expt}{Dataset containing experiment information}
}
\value{
A dataset containing the sensitivity of total biomass to temperature variation
}
\description{
Get the sensitivity of total biomass to temperature variation. Currently measured as the slope of a linear regression of total biomass on temperature.
}
\examples{
NULL
}
