% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_igr.R
\name{get_delta_igr}
\alias{get_delta_igr}
\title{Calculate the difference in growth rate from one time point to the next}
\usage{
get_delta_igr(experiment_folder, experiment_design_filename, every_t = 1)
}
\arguments{
\item{experiment_folder}{The folder where the experiment is stored}

\item{experiment_design_filename}{The name of the experiment design file}

\item{every_t}{The time interval at which to calculate the difference}
}
\value{
Nothing. Saves data to a file.
}
\description{
Calculate the difference in growth rate from one time point to the next
}
\examples{
NULL
}
