% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temporal_derivatives.R
\name{get_temporal_derivatives}
\alias{get_temporal_derivatives}
\title{Calculate the derivatives of the growth rate - temperature relationship at the temperatures in the temperature times series of each case in the experiment.}
\usage{
get_temporal_derivatives(
  experiment_folder,
  experiment_design_filename,
  every_t = 10
)
}
\arguments{
\item{experiment_folder}{The folder where the experiment is stored}

\item{experiment_design_filename}{The name of the experiment design file}

\item{every_t}{The time interval at which to calculate the derivatives}
}
\value{
Nothing. Saves data to a file.
}
\description{
Calculate the derivatives of the growth rate - temperature relationship at the temperatures in the temperature times series of each case in the experiment.
}
\examples{
NULL
}
