% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_with_inf_check.R
\name{lm_with_inf_check}
\alias{lm_with_inf_check}
\title{Linear model with infinite value check}
\usage{
lm_with_inf_check(formula, data)
}
\arguments{
\item{formula}{Linear regression model y ~ x}

\item{data}{Data frame containing the data}
}
\value{
A linear model object with slope replaced with NA if the data contains any Inf values
}
\description{
Linear model with infinite value check
}
\examples{
NULL
}
