% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_experiment_design_json.R
\name{read_experiment_design_json}
\alias{read_experiment_design_json}
\title{Read in the JSON formatted text file that contains the experiment design.
Most scalar values in the JSON file are expressions that can be evaluated
with \code{eval()} to get the values of the experiment design. Structured values,
such as list-based interaction specifications, are returned as-is.}
\usage{
read_experiment_design_json(experiment_folder, experiment_design_filename)
}
\arguments{
\item{experiment_folder}{The folder where the experiment information is located}

\item{experiment_design_filename}{The name of the file that contains the experiment design. This file should be in the experiment_folder and should be a JSON file. The JSON file should contain a list of expressions that can be evaluated to get the values of the experiment design.}
}
\value{
Returns a named list of expressions or structured values.
}
\description{
Read in the JSON formatted text file that contains the experiment design.
Most scalar values in the JSON file are expressions that can be evaluated
with \code{eval()} to get the values of the experiment design. Structured values,
such as list-based interaction specifications, are returned as-is.
}
\examples{
NULL
}
