% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_lv.R
\name{simulator_lv}
\alias{simulator_lv}
\title{Simulate the population dynamics of a community of species using the Lotka-Volterra competition model with temperature-dependent vital rates.}
\usage{
simulator_lv(input_com_params, TcelSeries, initial_abundances)
}
\arguments{
\item{input_com_params}{Community object, containing all species and community parameters}

\item{initial_abundances}{}
}
\description{
Simulate the population dynamics of a community of species using the Lotka-Volterra competition model with temperature-dependent vital rates.
}
