% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_environments.R
\name{create_environments}
\alias{create_environments}
\title{Create temperature times series}
\usage{
create_environments(experiment_folder, experiment_design_filename)
}
\arguments{
\item{experiment_folder}{The folder where all information about the experiment is stored}

\item{experiment_design_filename}{The filename of the experiment design file}
}
\value{
Returns nothing. Saves to SQLite databases the temperature time series.
}
\description{
Create temperature times series
}
\examples{
NULL
}
