% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_environments.R
\name{create_environments}
\alias{create_environments}
\title{Create temperature times series. Currently three options for how times series vary among cases.
\enumerate{
\item \code{same_per_replicate} replicates of the same number share the same environmental time series. E.g., \code{case1_rep1} and \code{case2_rep1} share the same time series.
\item \code{all_different} all environmental time series are different
}}
\usage{
create_environments(experiment_folder, experiment_design_filename)
}
\arguments{
\item{experiment_folder}{The folder where all information about the experiment is stored}

\item{experiment_design_filename}{The filename of the experiment design file}
}
\value{
Returns nothing. Saves to SQLite databases the temperature time series.
}
\description{
Create temperature times series. Currently three options for how times series vary among cases.
\enumerate{
\item \code{same_per_replicate} replicates of the same number share the same environmental time series. E.g., \code{case1_rep1} and \code{case2_rep1} share the same time series.
\item \code{all_different} all environmental time series are different
}
}
\examples{
NULL
}
