% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment_folder.R
\name{create_experiment_folder}
\alias{create_experiment_folder}
\title{Create folder for experiment}
\usage{
create_experiment_folder(experiment_folder_location, experiment_name)
}
\arguments{
\item{experiment_folder_location}{Location of experiment folder}

\item{experiment_name}{Name of experiment}
}
\value{
A message indicating if the folder was created or if it already exists, and the location of the folder
}
\description{
Create folder for experiment
}
\examples{
NULL
}
