% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment_table.R
\name{create_experiment_table}
\alias{create_experiment_table}
\title{This function takes the experimental design from the specified JSON experiment design file and creates a table of all the simulations that will be run.}
\usage{
create_experiment_table(
  experiment_folder,
  experiment_design_filename,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{experiment_folder}{Folder where the experiment data will be saved}

\item{experiment_design_filename}{Name of the experiment definition file}

\item{overwrite}{Logical. If \code{TRUE}, overwrite an existing experiment table.}

\item{verbose}{Logical. If \code{TRUE}, print messages about written outputs.}
}
\value{
Returns the number of cases in the experiment. Also saves to RDS the experiment design, for later use.
}
\description{
This function takes the experimental design from the specified JSON experiment design file and creates a table of all the simulations that will be run.
}
\details{
LV experiment definitions can use the preferred \code{lv_interactions}
field to specify one or more named interaction treatments. Each treatment
can set \code{type}, \code{symmetry}, \code{distribution}, \code{parameters}, and \code{diagonal}.
Legacy \verb{alpha_ij_*} fields are still supported and are converted to
interaction specifications internally.
}
\examples{
NULL
}
