% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment_table.R
\name{create_experiment_table}
\alias{create_experiment_table}
\title{This function takes the experimental design from the specified JSON experiment design file and creates a table of all the simulations that will be run.}
\usage{
create_experiment_table(experiment_folder, experiment_design_filename)
}
\arguments{
\item{experiment_folder}{Folder where the experiment data will be saved}

\item{experiment_design_filename}{Name of the experiment definition file}
}
\value{
Returns the number of cases in the experiment. Also saves to RDS the experiment design, for later use.
}
\description{
This function takes the experimental design from the specified JSON experiment design file and creates a table of all the simulations that will be run.
}
\examples{
NULL
}
