% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_CPC_measures.R
\name{get_community_CPC_measures}
\alias{get_community_CPC_measures}
\title{Get measures of community performance curves}
\usage{
get_community_CPC_measures(temperatures, expt, expt_def, every_t = 1)
}
\arguments{
\item{temperatures}{\itemize{
\item temperature time series used in the simulations
}}

\item{expt}{\itemize{
\item experiment table with community parameters
}}

\item{expt_def}{\itemize{
\item experiment design information
}}

\item{every_t}{\itemize{
\item how often to sample the temperature time series (e.g., every 1 time step, every 10 time steps, etc.)
}}
}
\value{
A dataset containing measures of community performance curves, including CV of community performance, synchrony of performance curves, and average CV of species performance curves weighted by their mean contribution to community performance.
}
\description{
Get measures of community performance curves
}
\examples{
NULL
}
