% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_measures.R
\name{get_community_measures}
\alias{get_community_measures}
\title{Get various community level measures, e.g., community stability, response diversity, position of optimal temperature, etc.}
\usage{
get_community_measures(experiment_folder, experiment_design_filename)
}
\arguments{
\item{experiment_folder}{The folder containing the experiment data}

\item{experiment_design_filename}{The name of the experiment design file}
}
\value{
Nothing. Saves data to a file.
}
\description{
Get various community level measures, e.g., community stability, response diversity, position of optimal temperature, etc.
}
\examples{
NULL
}
