% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_sum_derivatives.R
\name{get_community_sum_derivatives}
\alias{get_community_sum_derivatives}
\title{Get the sum of the communities derivatives at arbitrary temperatures and at actual (temp = temporal) temperatures}
\usage{
get_community_sum_derivatives(arb_derivs, temp_derivs)
}
\arguments{
\item{arb_derivs}{Connection to database containing arbitrary derivatives}

\item{temp_derivs}{Connection to database containing temporal (actual) derivatives}
}
\value{
A dataset containing the sum of the derivatives at arbitrary and actual temperatures
}
\description{
Get the sum of the communities derivatives at arbitrary temperatures and at actual (temp = temporal) temperatures
}
\examples{
NULL
}
