% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate_gaussian.R
\name{intrinsic_growth_gaussian}
\alias{intrinsic_growth_gaussian}
\title{Calculate intrinsic growth rate from species parameters and a temperature, assuming a Gaussian birth rate - temperature response curve and an exponential death rate - temperature response curve.}
\usage{
intrinsic_growth_gaussian(a_b_i, b_opt_i, s_i, a_d_i, z_i, temperature)
}
\arguments{
\item{a_b_i}{Value of birth rate when temperature is equal to b_opt_i}

\item{b_opt_i}{Temperature at which birth rate is maximized}

\item{s_i}{Spread of the Gaussian curve}

\item{a_d_i}{Value of death rate when temperature is equal to 0}

\item{z_i}{Slope of the exponential curve}

\item{temperature}{Temperature at which to calculate the intrinsic growth rate}
}
\value{
Returns the intrinsic growth rate at the given temperature
}
\description{
Calculate intrinsic growth rate from species parameters and a temperature, assuming a Gaussian birth rate - temperature response curve and an exponential death rate - temperature response curve.
}
\examples{
NULL
}
