% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate_gaussian.R
\name{intrinsic_growth_gaussian}
\alias{intrinsic_growth_gaussian}
\title{Calculate intrinsic growth rate from species parameters and temperature}
\usage{
intrinsic_growth_gaussian(
  birth_maximum,
  birth_optimum,
  birth_width,
  death_intercept,
  death_temperature_slope,
  temperature
)
}
\arguments{
\item{birth_maximum}{Value of birth rate at the birth optimum.}

\item{birth_optimum}{Temperature at which birth rate is maximized.}

\item{birth_width}{Standard-deviation width of the Gaussian birth-rate curve.}

\item{death_intercept}{Value of death rate when temperature is equal to 0.}

\item{death_temperature_slope}{Slope of the exponential death-rate curve.}

\item{temperature}{Temperature at which to calculate the intrinsic growth
rate.}
}
\value{
Returns the intrinsic growth rate at the given temperature
}
\description{
Assumes a Gaussian birth-rate temperature response curve and an exponential
death-rate temperature response curve.
}
\examples{
NULL
}
