% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_a_community.R
\name{make_a_community}
\alias{make_a_community}
\title{Make a community parameter object}
\usage{
make_a_community(
  S,
  a_b_mean,
  a_b_range,
  a_b_distribution,
  b_opt_mean,
  b_opt_range,
  b_opt_distribution,
  sd_perf_distribution,
  sd_perf_mean,
  sd_perf_range,
  alpha_ij_mean,
  alpha_ij_sd,
  community_seed,
  a_d,
  z,
  alpha_jj,
  alpha_ij_distribution
)
}
\arguments{
\item{S}{Number of species in the community}

\item{a_b_mean}{Mean of the distribution from which a_b values are drawn.}

\item{a_b_range}{Range of the distribution from which a_b values are drawn.}

\item{a_b_distribution}{Distribution used to generate \code{a_b_i} values.}

\item{b_opt_mean}{Mean of the distribution from which \code{b_opt_i} values are drawn.}

\item{b_opt_range}{Range of the distribution from which \code{b_opt_i} values are drawn.}

\item{b_opt_distribution}{Distribution used to generate \code{b_opt_i} values.}

\item{sd_perf_distribution}{Distribution used to generate Gaussian
performance-curve widths.}

\item{sd_perf_mean}{Mean of the distribution from which Gaussian
performance-curve widths are drawn.}

\item{sd_perf_range}{Range of the distribution from which Gaussian
performance-curve widths are drawn.}

\item{alpha_ij_mean}{Mean of the distribution from which the off-diagonal elements of the community matrix are drawn.}

\item{alpha_ij_sd}{Standard deviation of the distribution from which the off-diagonal elements of the community matrix are drawn.}

\item{community_seed}{Random seed used when generating community traits.}

\item{a_d}{Death rate when temperature is equal to 0; same for all species.}

\item{z}{Exponential rate of increase in death rate with temperature; same for all species.}

\item{alpha_jj}{Value of the diagonal of the community matrix, shared across species.}

\item{alpha_ij_distribution}{Distribution used to generate off-diagonal interaction coefficients.}
}
\value{
Returns a list containing the community object.
}
\description{
This function makes a community object containing the parameters for all
species in the community. At the moment, only the temperature optima
\code{b_opt_i} vary among species; the remaining parameters are shared or are
generated from common distributions.
}
\examples{
NULL
}
