% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_a_community.R
\name{make_a_community}
\alias{make_a_community}
\title{This function makes a community object, which contains all the parameters of the all speices in the community. At the moment only the temperature optima $b_{opt_i}$ can vary among the species. The other parameters are the same for all species.}
\usage{
make_a_community(
  S,
  a_b_mean,
  a_b_range,
  a_b_distribution,
  b_opt_mean,
  b_opt_range,
  b_opt_distribution,
  alpha_ij_mean,
  alpha_ij_sd,
  community_seed,
  s,
  a_d,
  z,
  alpha_jj
)
}
\arguments{
\item{S}{Number of species in the community}

\item{a_b_mean}{Mean of the distribution from which a_b values are drawn.}

\item{a_b_range}{Range of the distribution from which a_b values are drawn.}

\item{b_opt_mean}{Mean of the distribution from which b_opt_i values are drawn.}

\item{b_opt_range}{Range of the distribution from which b_opt_i values are drawn.}

\item{alpha_ij_mean}{Mean of the distribution from which the off-diagonal elements of the community matrix are drawn.}

\item{alpha_ij_sd}{Standard deviation of the distribution from which the off-diagonal elements of the community matrix are drawn.}

\item{s}{Standard deviation of the Gaussian birth rate - temperature response curve; same for all species.}

\item{a_d}{Death rate when temperature is equal to 0; same for all species.}

\item{z}{Exponential rate of increase in death rate with temperature; same for all species.}

\item{alpha_jj}{value of the diagonal of the community matrix, same for all jj}

\item{a_b}{Birth rate when temperature is equal to b_opt_i; same for all species.}

\item{trait_selection_method}{Either "random1" or "deterministic". If "random1", the b_opt_i values are drawn at random from a uniform distribution with mean b_opt_mean and range b_opt_range. If "deterministic", the b_opt_i values are equally spaced between b_opt_mean - 0.5\emph{b_opt_range and b_opt_mean + 0.5}b_opt_range.}
}
\value{
Returns a list containing the community object.
}
\description{
This function makes a community object, which contains all the parameters of the all speices in the community. At the moment only the temperature optima $b_{opt_i}$ can vary among the species. The other parameters are the same for all species.
}
\examples{
NULL
}
