% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_a_consumer_resource_community.R
\name{make_a_consumer_resource_community}
\alias{make_a_consumer_resource_community}
\title{Make a consumer-resource community parameter object}
\usage{
make_a_consumer_resource_community(
  S,
  u_max_mean,
  u_max_range,
  u_max_distribution,
  u_opt_mean,
  u_opt_range,
  u_opt_distribution,
  sd_u_mean,
  sd_u_range,
  sd_u_distribution,
  half_saturation_mean,
  half_saturation_range,
  half_saturation_distribution,
  consumer_death_rate,
  resource_renewal_rate,
  resource_supply,
  conversion_efficiency,
  resource_use_mode = "one_resource_all_consumers",
  active_resource = 1,
  resource_specialization = 1,
  resource_specialization_distribution = "constant",
  resource_specialization_mean = resource_specialization,
  resource_specialization_range = 0,
  resource_specialization_precision = 10,
  community_seed
)
}
\arguments{
\item{S}{Number of consumer species. The number of resources is set equal to
\code{S}, except when \code{resource_use_mode = "shared_to_private"}, where the
number of resources is \code{S + 1}.}

\item{u_max_mean}{Mean maximum uptake height.}

\item{u_max_range}{Range of maximum uptake heights.}

\item{u_max_distribution}{Distribution for maximum uptake heights.}

\item{u_opt_mean}{Mean thermal optimum for maximum uptake.}

\item{u_opt_range}{Range of thermal optima.}

\item{u_opt_distribution}{Distribution for thermal optima.}

\item{sd_u_mean}{Mean standard-deviation width of the Gaussian uptake curve.}

\item{sd_u_range}{Range of standard-deviation widths for Gaussian uptake
curves.}

\item{sd_u_distribution}{Distribution for Gaussian uptake-curve widths.}

\item{half_saturation_mean}{Mean Monod half-saturation constant.}

\item{half_saturation_range}{Range of half-saturation constants.}

\item{half_saturation_distribution}{Distribution for half-saturation constants.}

\item{consumer_death_rate}{Consumer death rate.}

\item{resource_renewal_rate}{Chemostat resource renewal rate.}

\item{resource_supply}{Resource supply concentration.}

\item{conversion_efficiency}{Conversion efficiency from uptake to consumer
growth.}

\item{resource_use_mode}{Resource-use mode. One of
\code{"one_resource_all_consumers"}, \code{"diagonal"}, or \code{"shared_to_private"}.}

\item{active_resource}{Active resource index for
\code{"one_resource_all_consumers"} or shared resource index for
\code{"shared_to_private"}.}

\item{resource_specialization}{Backwards-compatible scalar value between 0
and 1 controlling the transition from shared-resource use to private-
resource use when \code{resource_use_mode = "shared_to_private"}.}

\item{resource_specialization_distribution}{Distribution used to generate
species-level shared-private partition values. One of \code{"constant"},
\code{"regular"}, \code{"random_uniform"}, or \code{"beta"}.}

\item{resource_specialization_mean}{Mean species-level private-resource
specialization.}

\item{resource_specialization_range}{Range for \code{"regular"} and
\code{"random_uniform"} species-level private-resource specialization.}

\item{resource_specialization_precision}{Precision for beta-distributed
species-level private-resource specialization. Larger values produce less
among-species variation around \code{resource_specialization_mean}.}

\item{community_seed}{Random seed used when generating community traits.}
}
\value{
A consumer-resource community parameter object.
}
\description{
Make a consumer-resource community parameter object
}
\examples{
NULL
}
