% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_a_consumer_resource_community.R
\name{make_a_consumer_resource_community}
\alias{make_a_consumer_resource_community}
\title{Make a consumer-resource community parameter object}
\usage{
make_a_consumer_resource_community(
  S,
  u_max_mean,
  u_max_range,
  u_max_distribution,
  u_opt_mean,
  u_opt_range,
  u_opt_distribution,
  sd_u_mean,
  sd_u_range,
  sd_u_distribution,
  half_saturation_mean,
  half_saturation_range,
  half_saturation_distribution,
  consumer_death_rate,
  resource_renewal_rate,
  resource_supply,
  conversion_efficiency,
  resource_use_mode = "one_resource_all_consumers",
  active_resource = 1,
  resource_specialization = 1,
  community_seed
)
}
\arguments{
\item{S}{Number of consumer species. The number of resources is set equal to
\code{S}, except when \code{resource_use_mode = "shared_to_private"}, where the
number of resources is \code{S + 1}.}

\item{u_max_mean}{Mean maximum uptake height.}

\item{u_max_range}{Range of maximum uptake heights.}

\item{u_max_distribution}{Distribution for maximum uptake heights.}

\item{u_opt_mean}{Mean thermal optimum for maximum uptake.}

\item{u_opt_range}{Range of thermal optima.}

\item{u_opt_distribution}{Distribution for thermal optima.}

\item{sd_u_mean}{Mean Gaussian uptake breadth.}

\item{sd_u_range}{Range of Gaussian uptake breadths.}

\item{sd_u_distribution}{Distribution for Gaussian uptake breadths.}

\item{half_saturation_mean}{Mean Monod half-saturation constant.}

\item{half_saturation_range}{Range of half-saturation constants.}

\item{half_saturation_distribution}{Distribution for half-saturation constants.}

\item{consumer_death_rate}{Consumer death rate.}

\item{resource_renewal_rate}{Chemostat resource renewal rate.}

\item{resource_supply}{Resource supply concentration.}

\item{conversion_efficiency}{Conversion efficiency from uptake to consumer
growth.}

\item{resource_use_mode}{Resource-use mode. One of
\code{"one_resource_all_consumers"}, \code{"diagonal"}, or \code{"shared_to_private"}.}

\item{active_resource}{Active resource index for
\code{"one_resource_all_consumers"} or shared resource index for
\code{"shared_to_private"}.}

\item{resource_specialization}{Value between 0 and 1 controlling the
transition from shared-resource use to private-resource use when
\code{resource_use_mode = "shared_to_private"}.}

\item{community_seed}{Random seed used when generating community traits.}
}
\value{
A consumer-resource community parameter object.
}
\description{
Make a consumer-resource community parameter object
}
\examples{
NULL
}
