% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots_for_one_community.R
\name{make_plots_for_one_community}
\alias{make_plots_for_one_community}
\title{Make graphs showing variance results from one case}
\usage{
make_plots_for_one_community(experiment_folder, case_id_oi)
}
\arguments{
\item{experiment_folder}{The folder where the experiment is stored}

\item{case_id_oi}{The case id of the case to be plotted}
}
\value{
List of graphs.
}
\description{
Make graphs showing variance results from one case
}
\examples{
NULL
}
