% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dynamics.R
\name{simulate_dynamics}
\alias{simulate_dynamics}
\title{Simulate the dynamics of all the cases in an experiment
Unfortunately at the moment has features of temperature series hard coded in}
\usage{
simulate_dynamics(experiment_folder, experiment_design_filename)
}
\arguments{
\item{experiment_folder}{The folder where all information about the experiment is stored}

\item{experiment_design_filename}{The filename of the experiment design file}
}
\value{
Returns nothing. Saves to SQLite databases the temperature time series and the community dynamics.
}
\description{
Simulate the dynamics of all the cases in an experiment
Unfortunately at the moment has features of temperature series hard coded in
}
\examples{
NULL
}
