% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_lv_continuous.R
\name{simulator_lv_continuous}
\alias{simulator_lv_continuous}
\title{Simulate Lotka-Volterra community dynamics in continuous time}
\usage{
simulator_lv_continuous(
  input_com_params,
  TcelSeries,
  initial_abundances,
  times = seq_len(ncol(TcelSeries)),
  output_times = times,
  temperature_interpolation = "linear",
  immigration_rate = 0.1,
  immigration_mode = "continuous",
  ode_method = "lsoda",
  rtol = 1e-06,
  atol = 1e-08,
  max_step = 1,
  blowup_threshold = 1e+12
)
}
\arguments{
\item{input_com_params}{Community object, containing all species and
community parameters.}

\item{TcelSeries}{Time series of temperature values.}

\item{initial_abundances}{Initial abundances of each species.}

\item{times}{Times at which temperatures are defined. Defaults to one time
unit per temperature value.}

\item{output_times}{Times at which abundances should be returned.}

\item{temperature_interpolation}{How temperature should be interpolated
between supplied time points. One of \code{"linear"} or \code{"constant"}.}

\item{immigration_rate}{Immigration rate per species.}

\item{immigration_mode}{One of \code{"continuous"} or \code{"pulse"}.}

\item{ode_method}{ODE solver method passed to \code{deSolve::ode()}.}

\item{rtol}{Relative tolerance passed to \code{deSolve::ode()}.}

\item{atol}{Absolute tolerance passed to \code{deSolve::ode()}.}

\item{max_step}{Maximum solver step size.}

\item{blowup_threshold}{Abundance threshold above which the run stops.}
}
\value{
Time series of population abundances for each species.
}
\description{
Simulate Lotka-Volterra community dynamics in continuous time
}
\examples{
NULL
}
